+++ larva/src/main/java/org/frankframework/larva/LarvaTool.java
@@ -80,8 +80,8 @@ import org.frankframework.configuration.ClassNameRewriter;
 import org.frankframework.configuration.IbisContext;
 import org.frankframework.core.SenderException;
 import org.frankframework.core.TimeoutException;
-import org.frankframework.larva.queues.LarvaActionFactory;
-import org.frankframework.larva.queues.LarvaScenarioAction;
+import org.frankframework.larva.actions.LarvaActionFactory;
+import org.frankframework.larva.actions.LarvaScenarioAction;
 import org.frankframework.lifecycle.FrankApplicationInitializer;
 import org.frankframework.stream.FileMessage;
 import org.frankframework.stream.Message;
