+++ larva/src/main/java/org/frankframework/larva/actions/AbstractLarvaAction.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package org.frankframework.larva.queues;
+package org.frankframework.larva.actions;
 
 import java.util.Map;
 import java.util.Properties;
